+++ oap-server/analyzer/meter-analyzer/src/main/java/org/apache/skywalking/oap/meter/analyzer/Analyzer.java
@@ -42,11 +42,12 @@ import org.apache.skywalking.oap.meter.analyzer.dsl.FilterExpression;
 import org.apache.skywalking.oap.meter.analyzer.dsl.Result;
 import org.apache.skywalking.oap.meter.analyzer.dsl.Sample;
 import org.apache.skywalking.oap.meter.analyzer.dsl.SampleFamily;
-import org.apache.skywalking.oap.meter.analyzer.k8s.K8sInfoRegistry;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
 import org.apache.skywalking.oap.server.core.analysis.manual.endpoint.EndpointTraffic;
 import org.apache.skywalking.oap.server.core.analysis.manual.instance.InstanceTraffic;
+import org.apache.skywalking.oap.server.core.analysis.manual.relation.process.ProcessRelationClientSideMetrics;
+import org.apache.skywalking.oap.server.core.analysis.manual.relation.process.ProcessRelationServerSideMetrics;
 import org.apache.skywalking.oap.server.core.analysis.manual.relation.service.ServiceRelationClientSideMetrics;
 import org.apache.skywalking.oap.server.core.analysis.manual.relation.service.ServiceRelationServerSideMetrics;
 import org.apache.skywalking.oap.server.core.analysis.manual.service.ServiceTraffic;
@@ -241,10 +242,6 @@ public class Analyzer {
             }
         }
         createMetric(ctx.getScopeType(), metricType.literal, ctx.getDownsampling());
-
-        if (ctx.isRetagByK8sMeta()) {
-            K8sInfoRegistry.getInstance().start();
-        }
     }
 
     private void createMetric(final ScopeType scopeType,
@@ -262,16 +259,12 @@ public class Analyzer {
 
     private void generateTraffic(MeterEntity entity) {
         if (entity.getDetectPoint() != null) {
-            switch (entity.getDetectPoint()) {
-                case SERVER:
-                    entity.setServiceName(entity.getDestServiceName());
-                    toService(requireNonNull(entity.getDestServiceName()), entity.getLayer());
-                    serverSide(entity);
+            switch (entity.getScopeType()) {
+                case SERVICE_RELATION:
+                    serviceRelationTraffic(entity);
                     break;
-                case CLIENT:
-                    entity.setServiceName(entity.getSourceServiceName());
-                    toService(requireNonNull(entity.getSourceServiceName()), entity.getLayer());
-                    clientSide(entity);
+                case PROCESS_RELATION:
+                    processRelationTraffic(entity);
                     break;
                 default:
             }
@@ -309,7 +302,23 @@ public class Analyzer {
         MetricsStreamProcessor.getInstance().in(s);
     }
 
-    private void serverSide(MeterEntity entity) {
+    private void serviceRelationTraffic(MeterEntity entity) {
+        switch (entity.getDetectPoint()) {
+            case SERVER:
+                entity.setServiceName(entity.getDestServiceName());
+                toService(requireNonNull(entity.getDestServiceName()), entity.getLayer());
+                serviceRelationServerSide(entity);
+                break;
+            case CLIENT:
+                entity.setServiceName(entity.getSourceServiceName());
+                toService(requireNonNull(entity.getSourceServiceName()), entity.getLayer());
+                serviceRelationClientSide(entity);
+                break;
+            default:
+        }
+    }
+
+    private void serviceRelationServerSide(MeterEntity entity) {
         ServiceRelationServerSideMetrics metrics = new ServiceRelationServerSideMetrics();
         metrics.setTimeBucket(TimeBucket.getMinuteTimeBucket(System.currentTimeMillis()));
         metrics.setSourceServiceId(entity.sourceServiceId());
@@ -319,7 +328,7 @@ public class Analyzer {
         MetricsStreamProcessor.getInstance().in(metrics);
     }
 
-    private void clientSide(MeterEntity entity) {
+    private void serviceRelationClientSide(MeterEntity entity) {
         ServiceRelationClientSideMetrics metrics = new ServiceRelationClientSideMetrics();
         metrics.setTimeBucket(TimeBucket.getMinuteTimeBucket(System.currentTimeMillis()));
         metrics.setSourceServiceId(entity.sourceServiceId());
@@ -328,4 +337,37 @@ public class Analyzer {
         metrics.setEntityId(entity.id());
         MetricsStreamProcessor.getInstance().in(metrics);
     }
+
+    private void processRelationTraffic(MeterEntity entity) {
+        switch (entity.getDetectPoint()) {
+            case SERVER:
+                processRelationServerSide(entity);
+                break;
+            case CLIENT:
+                processRelationClientSide(entity);
+                break;
+            default:
+        }
+    }
+
+    private void processRelationServerSide(MeterEntity entity) {
+        ProcessRelationServerSideMetrics metrics = new ProcessRelationServerSideMetrics();
+        metrics.setTimeBucket(TimeBucket.getMinuteTimeBucket(System.currentTimeMillis()));
+        metrics.setServiceInstanceId(entity.serviceInstanceId());
+        metrics.setSourceProcessId(entity.getSourceProcessId());
+        metrics.setDestProcessId(entity.getDestProcessId());
+        metrics.setEntityId(entity.id());
+        MetricsStreamProcessor.getInstance().in(metrics);
+    }
+
+    private void processRelationClientSide(MeterEntity entity) {
+        ProcessRelationClientSideMetrics metrics = new ProcessRelationClientSideMetrics();
+        metrics.setTimeBucket(TimeBucket.getMinuteTimeBucket(System.currentTimeMillis()));
+        metrics.setServiceInstanceId(entity.serviceInstanceId());
+        metrics.setSourceProcessId(entity.getSourceProcessId());
+        metrics.setDestProcessId(entity.getDestProcessId());
+        metrics.setEntityId(entity.id());
+        MetricsStreamProcessor.getInstance().in(metrics);
+    }
+
 }