+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/IDManager.java
@@ -268,6 +268,32 @@ public class IDManager {
             return Hashing.sha256().newHasher().putString(String.format("%s_%s",
                     name, instanceId), Charsets.UTF_8).hash().toString();
         }
+
+        /**
+         * @return encoded process relation id
+         */
+        public static String buildRelationId(ProcessRelationDefine define) {
+            return define.sourceId + Const.RELATION_ID_CONNECTOR + define.destId;
+        }
+
+        /**
+         * @return process relation ID object decoded from {@link #buildRelationId(ProcessRelationDefine)} result
+         */
+        public static ProcessRelationDefine analysisRelationId(String entityId) {
+            String[] parts = entityId.split(Const.RELATION_ID_PARSER_SPLIT);
+            if (parts.length != 2) {
+                throw new RuntimeException("Illegal Process Relation entity id");
+            }
+            return new ProcessRelationDefine(parts[0], parts[1]);
+        }
+
+        @RequiredArgsConstructor
+        @Getter
+        @EqualsAndHashCode
+        public static class ProcessRelationDefine {
+            private final String sourceId;
+            private final String destId;
+        }
     }
 
     /**